Oh boy off I go changing EVERYTHING

+++ Assets/Scripts/Game/Battle/ScriptableObjects/CharacterStats.cs
@@ -5,12 +5,12 @@ namespace Game.Battle.ScriptableObjects
     [CreateAssetMenu(fileName = "CharacterStats", menuName = "Scriptable Objects/CharacterStats")]
     public class CharacterStats : ScriptableObject
     {
-        [SerializeField] private float maxStamina;
-        [SerializeField] private float maxHealth;
+        [SerializeField] private int maxStamina;
+        [SerializeField] private int maxHealth;
         
         //Simple getter variables, you should never modify a scriptable object
-        public float MaxStamina => maxStamina;
-        public float MaxHealth => maxHealth;
+        public int MaxStamina => maxStamina;
+        public int MaxHealth => maxHealth;
         
     }
 }
